Move test functions near struct and usage
